PDF/backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
transformers>=4.43.2
onnxruntime>=1.18.1
tokenizers>=0.19.1
huggingface-hub>=0.30.0
